requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
fastapi==0.135.1
filelock==3.25.2
fsspec==2026.2.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==1.26.4
pandas==3.0.1
pillow==12.1.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-multipart==0.0.22
six==1.17.0
starlette==0.52.1
sympy==1.14.0
torch==2.2.2
torchvision==0.17.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
